bots/rockfish-ai/requirements.txt
python-socketio[client]>=5.12,<5.13